pyproject.toml
[project]
name = "swing-trader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "pandas-ta>=0.4.71b0",
    "plotly>=6.7.0",
    "pydantic>=2.13.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "typer>=0.24.1",
    "yfinance>=1.2.2",
]

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "ruff>=0.15.10",
]
